Moved the maybe_run_functions_eagerly to test_utils.
